+++ src/common/NetworkCommon.hpp
@@ -13,7 +13,6 @@ class NetworkResult;
 
 using NetworkSuccessCallback = std::function<void(NetworkResult)>;
 using NetworkErrorCallback = std::function<void(NetworkResult)>;
-using NetworkReplyCreatedCallback = std::function<void(QNetworkReply *)>;
 using NetworkFinallyCallback = std::function<void()>;
 
 enum class NetworkRequestType {
@@ -22,13 +21,6 @@ enum class NetworkRequestType {
     Put,
     Delete,
     Patch,
-};
-const static std::vector<QString> networkRequestTypes{
-    "GET",     //
-    "POST",    //
-    "PUT",     //
-    "DELETE",  //
-    "PATCH",   //
 };
 
 // parseHeaderList takes a list of headers in string form,